requirements.txt
flask==0.9
flask-login
sqlalchemy==0.7.9
flask-sqlalchemy==0.16
sqlalchemy-migrate==0.7.2
flup
flask-markdown
flask-testing
flask-wtf
coverage
nose
